Finishing touches to mesh-cutting
