+++ ChiTech/ChiMesh/MeshCutting/meshcutting.h
@@ -35,9 +35,9 @@ namespace mesh_cutting
   };
 
   //2D_utils
-  std::pair<uint64_t,uint64_t>
-    MakeEdgeFromPolygonEdgeIndex(const std::vector<uint64_t>& vertex_ids,
-                                 size_t edge_index);
+  Edge MakeUniqueEdge(const Edge& edge);
+  Edge MakeEdgeFromPolygonEdgeIndex(const std::vector<uint64_t>& vertex_ids,
+                                    size_t edge_index);
   chi_mesh::Vector3 GetEdgeCentroid(const Edge& edge,
                                     const chi_mesh::MeshContinuum& grid);
 
@@ -46,11 +46,11 @@ namespace mesh_cutting
     const std::vector<uint64_t>& vertex_ids,
     chi_mesh::Cell& cell);
 
-  bool CheckPolygonQuality(const MeshContinuum& mesh, const chi_mesh::Cell& cell);
+  bool CheckPolygonQuality(const MeshContinuum& mesh,
+                           const chi_mesh::Cell& cell,
+                           bool check_convexity=false);
 
-  void SplitConcavePolygonsIntoTriangles(MeshContinuum& mesh,
-                                         std::vector<chi_mesh::Cell*>& cell_list);
-
+  //2D_cutcell
   void CutPolygon(const std::vector<ECI>& cut_edges,
                   const std::set<uint64_t>& cut_vertices,
                   const Vector3 &plane_point,
@@ -60,8 +60,12 @@ namespace mesh_cutting
 
   //3D_utils
   bool CheckPolyhedronQuality(const MeshContinuum& mesh,
-                              const chi_mesh::Cell& cell);
+                              const chi_mesh::Cell& cell,
+                              bool check_convexity=false);
 
+  std::set<size_t> FindNeighborFaceIndices(
+    const std::vector<std::vector<uint64_t>>& proxy_faces,
+    size_t face_index);
 
   std::vector<Edge> FindNonManifoldEdges(
     const std::vector<std::vector<uint64_t>>& proxy_faces);
